Revert "[WIP] Refactor code for clearer separation of robot and vision components"
