+++ src/biemhTekniker/Main.java
@@ -13,8 +13,8 @@ import biemhTekniker.managers.PLCManager;
 import biemhTekniker.programs.RobotDispatcher;
 import biemhTekniker.programs.VisionDispatcher;
 import biemhTekniker.programs.ProgramRange;
-import biemhTekniker.lib.robot.RobotContext;
-import biemhTekniker.lib.vision.VisionContext;
+import biemhTekniker.programs.RobotContext;
+import biemhTekniker.programs.VisionContext;
 import biemhTekniker.vision.SmartPickingThread;
 import biemhTekniker.vision.VisionManager;
 import com.kuka.common.ThreadUtil;
