+++ src/biemhTekniker/programs/IndividualVisionCommandTask.java
@@ -1,7 +1,4 @@
-package biemhTekniker.programs.vision;
-
-import biemhTekniker.lib.vision.VisionTask;
-import biemhTekniker.lib.vision.VisionContext;
+package biemhTekniker.programs;
 
 import biemhTekniker.logger.Logger;
 import biemhTekniker.vision.SmartPickingProtocol;
