+++ src/biemhTekniker/programs/PlaceNewWorkpieceProgram.java
@@ -1,8 +1,4 @@
-package biemhTekniker.programs.robot;
-
-import biemhTekniker.lib.robot.RobotProgram;
-import biemhTekniker.lib.robot.RobotContext;
-import biemhTekniker.lib.motion.*;
+package biemhTekniker.programs;
 
 import biemhTekniker.exceptions.ProgramCancelledException;
 import biemhTekniker.logger.Logger;
