+++ src/biemhTekniker/programs/RobotDispatcher.java
@@ -1,8 +1,5 @@
 package biemhTekniker.programs;
 
-import biemhTekniker.lib.robot.RobotContext;
-import biemhTekniker.lib.robot.RobotProgram;
-import biemhTekniker.programs.robot.*;
 import biemhTekniker.logger.Logger;
 import biemhTekniker.vision.SmartPickingThread;
