+++ src/biemhTekniker/programs/TestCalibrationProgram.java
@@ -1,8 +1,4 @@
-package biemhTekniker.programs.robot;
-
-import biemhTekniker.lib.robot.RobotProgram;
-import biemhTekniker.lib.robot.RobotContext;
-import biemhTekniker.lib.motion.*;
+package biemhTekniker.programs;
 
 import biemhTekniker.logger.Logger;
 import biemhTekniker.vision.SmartPickingProtocol;
